Add remaining re-linting changes (part 2 of 2)
The remaining re-linting changes pertain to new linting codes:

  * TC006/runtime-cast-value (formerly TCH006) suggests that cast calls
    use stringified type declarations.  Implemented.

  * B033/duplicate-value requires sets not to contain duplicate values.
    This is used in the `_types.js_truthiness` function to both test
    against and to document all of the falsy values except NaN.
    Ignored.

  * RUF043/pytest-raises-ambiguous-pattern requires patterns in the
    `match` argument of `pytest.raises` to be either raw strings, or
    passed through `re.escape`.  Implemented.

  * RUF046/unnecessary-cast-to-int forbids explicitly casting the result
    of `math.ceil` to int.  I believe `mypy`/`typeshed` originally had
    problems with this, but this no longer appears to be the case.
    Implemented.

+++ src/derivepassphrase/vault.py
@@ -253,7 +253,7 @@ class Vault:
             raise ValueError(msg)
         # Ensure the bound is strictly positive.
         entropy_bound = max(1, self._entropy())
-        return int(math.ceil(safety_factor * entropy_bound / 8))
+        return math.ceil(safety_factor * entropy_bound / 8)
 
     @staticmethod
     def _get_binary_string(s: bytes | bytearray | str, /) -> bytes:
